# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * GameZone API - gaming platform where users play games and earn rewards
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Game {
  id: number;
  title: string;
  description: string;
  genre: string;
  thumbnailUrl: string;
  gameUrl: string;
  creatorName: string;
  playCount: number;
  rating: number;
  rewardPerMinute: number;
  createdAt: string;
}

export interface GameInput {
  /** @minLength 1 */
  title: string;
  description: string;
  genre: string;
  thumbnailUrl: string;
  gameUrl: string;
  creatorName: string;
  rewardPerMinute: number;
}

export interface GameUpdate {
  title?: string;
  description?: string;
  genre?: string;
  thumbnailUrl?: string;
  gameUrl?: string;
  rewardPerMinute?: number;
}

export interface User {
  id: number;
  username: string;
  email: string;
  avatarUrl: string;
  balance: number;
  totalEarnings: number;
  gamesPlayed: number;
  createdAt: string;
}

export interface UserInput {
  /** @minLength 1 */
  username: string;
  email: string;
  /** @nullable */
  avatarUrl?: string | null;
}

export interface PlaySession {
  id: number;
  userId: number;
  gameId: number;
  /** @nullable */
  gameName?: string | null;
  status: string;
  pointsEarned: number;
  /** @nullable */
  durationMinutes?: number | null;
  startedAt: string;
  /** @nullable */
  endedAt?: string | null;
}

export interface UserStats {
  userId: number;
  totalPlayTime: number;
  totalEarnings: number;
  gamesPlayed: number;
  /** @nullable */
  favoriteGenre: string | null;
  recentSessions: PlaySession[];
}

export interface PlaySessionInput {
  userId: number;
  gameId: number;
}

export interface PlaySessionEnd {
  durationMinutes: number;
}

export interface LeaderboardEntry {
  rank: number;
  userId: number;
  username: string;
  avatarUrl: string;
  totalEarnings: number;
  gamesPlayed: number;
}

export interface Earning {
  id: number;
  userId: number;
  amount: number;
  type: string;
  status: string;
  createdAt: string;
}

export interface WithdrawalInput {
  userId: number;
  amount: number;
}

export type DashboardStatsTopGenresItem = {
  genre: string;
  count: number;
};

export interface DashboardStats {
  totalGames: number;
  totalPlayers: number;
  totalPayouts: number;
  activeSessionsCount: number;
  topGenres: DashboardStatsTopGenresItem[];
}

export type ListGamesParams = {
genre?: string;
search?: string;
};

export type ListPlaySessionsParams = {
userId?: number;
gameId?: number;
};

export type ListEarningsParams = {
userId?: number;
};