# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * GameZone API - gaming platform where users play games and earn rewards
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardStatsTopGenresItem } from './dashboardStatsTopGenresItem';

export interface DashboardStats {
  totalGames: number;
  totalPlayers: number;
  totalPayouts: number;
  activeSessionsCount: number;
  topGenres: DashboardStatsTopGenresItem[];
}