# lib/api-zod/src/generated/types/gameUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * GameZone API - gaming platform where users play games and earn rewards
 * OpenAPI spec version: 0.1.0
 */

export interface GameUpdate {
  title?: string;
  description?: string;
  genre?: string;
  thumbnailUrl?: string;
  gameUrl?: string;
  rewardPerMinute?: number;
}