# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * GameZone API - gaming platform where users play games and earn rewards
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  username: string;
  email: string;
  avatarUrl: string;
  balance: number;
  totalEarnings: number;
  gamesPlayed: number;
  createdAt: string;
}